refactor: remove unnecessary static cast

+++ src/tachimawari/control/controller/module/cm740.cpp
@@ -253,8 +253,7 @@ bool CM740::bulk_read_packet()
 
     if (ping(packet::protocol_1::PacketId::CONTROLLER)) {
       instruction_packet.add(
-        packet::protocol_1::PacketId::CONTROLLER, CM740Address::DXL_POWER,
-        static_cast<uint8_t>(30));
+        packet::protocol_1::PacketId::CONTROLLER, CM740Address::DXL_POWER, 30u);
     }
 
     if (instruction_packet.is_parameters_filled()) {
